Move editor schema logic from adapter into editors.ts
