--- packages/tux-adapter-contentful/src/widget-to-editor.tsx
@@ -1,19 +1,0 @@
-import {
-  BooleanField,
-  DatePicker,
-  Dropdown,
-  ImageField,
-  MarkdownField,
-  Radio,
-  TextField,
-} from 'tux'
-
-export default {
-  assetLinkEditor: ImageField,
-  boolean: BooleanField,
-  datePicker: DatePicker,
-  dropdown: Dropdown,
-  markdown: MarkdownField,
-  radio: Radio,
-  singleLine: TextField,
-}
